requirements.txt
streamlit
pdf2image
python-dotenv
google-generativeai
langchain-openai
langchain
openai
PyPDF2